requirements.txt
discord.py==2.0.1
pymongo[srv]==4.2.0
requests
BeautifulSoup4